+++ src/containers/DafaultAside.vue
@@ -117,7 +117,7 @@
           <div class="py-3 pb-5 mr-3 float-left">
             <div class="avatar">
               <img src="static/img/avatars/7.jpg" class="img-avatar" alt="admin@bootstrapmaster.com">
-              <b-badge variant="success" class="avatar-status"></b-badge>
+              <b-badge variant="danger" class="avatar-status"></b-badge>
             </div>
           </div>
           <div>
@@ -132,7 +132,7 @@
           <div class="py-3 pb-5 mr-3 float-left">
             <div class="avatar">
               <img src="static/img/avatars/7.jpg" class="img-avatar" alt="admin@bootstrapmaster.com">
-              <b-badge variant="success" class="avatar-status"></b-badge>
+              <b-badge variant="info" class="avatar-status"></b-badge>
             </div>
           </div>
           <div>
@@ -147,7 +147,7 @@
           <div class="py-3 pb-5 mr-3 float-left">
             <div class="avatar">
               <img src="static/img/avatars/7.jpg" class="img-avatar" alt="admin@bootstrapmaster.com">
-              <b-badge variant="success" class="avatar-status"></b-badge>
+              <b-badge variant="warning" class="avatar-status"></b-badge>
             </div>
           </div>
           <div>
@@ -162,7 +162,7 @@
           <div class="py-3 pb-5 mr-3 float-left">
             <div class="avatar">
               <img src="static/img/avatars/7.jpg" class="img-avatar" alt="admin@bootstrapmaster.com">
-              <b-badge variant="success" class="avatar-status"></b-badge>
+              <b-badge variant="dark" class="avatar-status"></b-badge>
             </div>
           </div>
           <div>